requirements.txt
langchain
langgraph
python-dotenv
langchain-community
huggingface_hub
ollama
langchain-openai
langchain-ollama
langchain-core
langchain_huggingface
streamlit
pypdf
beautifulsoup4
sympy
wikipedia